% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barriers_plot.R
\name{barriers_plot}
\alias{barriers_plot}
\title{Barriers Plot}
\usage{
barriers_plot(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A .png image of the plot
}
\description{
Generate the plot of resident barriers to healthy food
}
\examples{
barriers_plot(d)
}
