% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_demos.R
\name{basic_demos}
\alias{basic_demos}
\title{Basic SAFE demographics}
\usage{
basic_demos(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A list with three datasets
}
\description{
Calculate basic demographics information from the SAFE survey responses. Outputs a list with three datasets: the number of surveys completed per person, the number of kids for each household, and a table of summaries
}
\examples{
demos_list <- basic_demos(d)
}
