% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_insecurity_plot.R
\name{food_insecurity_plot}
\alias{food_insecurity_plot}
\title{Food insecurity plot}
\usage{
food_insecurity_plot(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A .png image of the plot
}
\description{
Generates the plot of food insecurity over time
}
\examples{
food_insecurity_plot(d)
}
