% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighborhoods.R
\name{get_neighborhoods}
\alias{get_neighborhoods}
\title{Assign Neighborhoods}
\usage{
get_neighborhoods(d, d_tract)
}
\arguments{
\item{d}{Your redcap dataset with record_id's}

\item{d_tract}{Your geocoded to the tract level dataset}
}
\value{
Your redcap dataset with neighborhoods attached
}
\description{
Insert your redcap data and the output from geocoding to the tract level
}
\examples{
d <- get_neighborhoods(d, d_tract)

}
