% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_ids.R
\name{get_valid_ids}
\alias{get_valid_ids}
\title{Filter to valid responses}
\usage{
get_valid_ids(d)
}
\arguments{
\item{d}{Your raw redcap dataset}
}
\value{
Your validated dataset
}
\description{
Filters your redcap dataset to only valid record IDs. Also eliminates rows with a duration less than 15 seconds and where respondents completed both language surveys
}
\examples{
d <- get_valid_ids(d)

}
