% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_card_list.R
\name{gift_card_list}
\alias{gift_card_list}
\title{Monthly Gift Card List}
\usage{
gift_card_list(url, token)
}
\arguments{
\item{url}{Your redcap project url}

\item{token}{Your personal redcap project token}
}
\value{

}
\description{
Input your REDCap project url and token and generate a csv file containing all eligible gift card recipients for the past month
}
\examples{
gift_card_list(my_url, my_token)

}
