% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_plot.R
\name{likert_plot}
\alias{likert_plot}
\title{Likert (Agree-Disagree) Plot}
\usage{
likert_plot(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood. Needs the columns to retain their original order}
}
\value{
A .png image of the plot
}
\description{
Generates a likert plot of the agree-disagree spectrum questions
}
\examples{
likert_plot(d)
}
