% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needs_plot.R
\name{needs_plot}
\alias{needs_plot}
\title{Resident needs plot}
\usage{
needs_plot(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A .png image of the plot
}
\description{
Generates the plot breaking down the resident needs
}
\examples{
needs_plot(d)
}
