% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_totals.R
\name{race_totals}
\alias{race_totals}
\title{Respondent summary by race}
\usage{
race_totals(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A dataframe with the number of responses by zip code
}
\description{
Respondent summary by race
}
\examples{
d_race <- race_totals(d)
}
