% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thriving_plot.R
\name{thriving_plot}
\alias{thriving_plot}
\title{Thriving plot}
\usage{
thriving_plot(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A .png image of the plot
}
\description{
Generates the plot of thriving over time
}
\examples{
thriving_plot(d)

}
