% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_totals.R
\name{zip_totals}
\alias{zip_totals}
\title{Respondent summary by zip code}
\usage{
zip_totals(d)
}
\arguments{
\item{d}{Your dataset geocoded to neighborhood}
}
\value{
A dataframe with the number of responses by zip code
}
\description{
Calculate the number of completed responses by zip code
}
\examples{
d_zip <- zip_totals(d)
}
